time: refactoring
Refactor time::instant::coarse to move atomic type into its own
module. This makes it consistent with the other modules.
